+++ test/integration/braintree/transaction_gateway_spec.js
@@ -6279,4 +6279,151 @@ describe('TransactionGateway', function () {
       });
     });
   });
+
+  describe('Multi Auth Adjustment endpoint on successful', function () {
+    it('returns success response', function (done) {
+      let transactionParams = {
+        merchantAccountId: specHelper.fakeFirstDataMerchantAccountId,
+        amount: '75.50',
+        creditCard: {
+          number: '5555555555554444',
+          expirationDate: '05/12'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '85.00', function (err, response) {
+          assert.isTrue(response.success);
+          assert.equal(response.transaction.amount, '85.00');
+          done();
+        })
+      );
+    });
+  });
+
+  describe('Multi Auth Adjustment endpoint on failure', function () {
+    let firstDataMasterTransactionParams = {
+      merchantAccountId: specHelper.fakeFirstDataMerchantAccountId,
+      amount: '75.50',
+      creditCard: {
+        number: '5555555555554444',
+        expirationDate: '06/09'
+      }
+    };
+
+    it('returns error code, when processor does not support multi auth adjustment', function (done) {
+      let transactionParams = {
+        merchantAccountId: specHelper.defaultMerchantAccountId,
+        amount: '75.50',
+        creditCard: {
+          number: CreditCardNumbers.CardTypeIndicators.Visa,
+          expirationDate: '06/09'
+        }
+      };
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '75.50', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('transaction').on('base')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.ProcessorDoesNotSupportAuthAdjustment);
+          done();
+        })
+      );
+    });
+
+    it('returns error code, when submitted adjustment amount is zero', function (done) {
+      specHelper.defaultGateway.transaction.sale(firstDataMasterTransactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '0.0', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('authorizationAdjustment').on('amount')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.AdjustmentAmountMustBeGreaterThanZero);
+          done();
+        })
+      );
+    });
+
+    it('returns error code, when adjusted amount is same as authorized amount', function (done) {
+      specHelper.defaultGateway.transaction.sale(firstDataMasterTransactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '75.50', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('authorizationAdjustment').on('base')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.NoNetAmountToPerformAuthAdjustment);
+          done();
+        })
+      );
+    });
+
+    it('returns error code, when transaction status is not authorized', function (done) {
+      let additionalParams = {options: {submitForSettlement: true}};
+      let transactionParams = Object.assign(additionalParams, firstDataMasterTransactionParams);
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '75.50', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('transaction').on('base')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.TransactionMustBeInStateAuthorized);
+          done();
+        })
+      );
+    });
+
+    it('returns error code, when transaction type is undefined or final', function (done) {
+      let additionalParams = {transactionSource: 'recurring_first'};
+      let transactionParams = Object.assign(additionalParams, firstDataMasterTransactionParams);
+
+      specHelper.defaultGateway.transaction.sale(transactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '85.50', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('transaction').on('base')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.TransactionIsNotEligibleForAdjustment);
+          done();
+        })
+      );
+    });
+
+    let firstDataVisaTransactionParams = {
+      merchantAccountId: specHelper.fakeFirstDataMerchantAccountId,
+      amount: '75.50',
+      creditCard: {
+        number: CreditCardNumbers.CardTypeIndicators.Visa,
+        expirationDate: '06/09'
+      }
+    };
+
+    it('returns error code, when processor does not support incremental auth', function (done) {
+      specHelper.defaultGateway.transaction.sale(firstDataVisaTransactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '85.50', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('transaction').on('base')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.ProcessorDoesNotSupportIncrementalAuth);
+          done();
+        })
+      );
+    });
+
+    it('returns error code, when processor does not support auth reversal', function (done) {
+      specHelper.defaultGateway.transaction.sale(firstDataVisaTransactionParams, (err, response) =>
+        specHelper.defaultGateway.transaction.adjustAuthorization(response.transaction.id, '65.50', function (err, response) {
+          assert.isFalse(response.success);
+          assert.equal(response.transaction.amount, '75.50');
+          let errorCode = response.errors.for('transaction').on('base')[0].code;
+
+          assert.equal(errorCode, ValidationErrorCodes.Transaction.ProcessorDoesNotSupportPartialAuthReversal);
+          done();
+        })
+      );
+    });
+  });
 });