+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## unreleased
+* Support all options for `gateway.transaction.submitForSettlement` and `gateway.transaction.submitForPartialSettlement`
+* Fix issue where options passed to `gateway.transaction.submitForSettlement` were being ignored (closes #168)
+* Fix issue where options passed to `gateway.transaction.submitForPartialSettlement` were being ignored
+
 ## 3.2.2
 - Update `dateformat` module to v4.5.1