+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.8.0
 * Add `LocalPaymentReversed` webhook notification support
 * Support all options for `gateway.transaction.submitForSettlement` and `gateway.transaction.submitForPartialSettlement`
 * Fix issue where options passed to `gateway.transaction.submitForSettlement` were being ignored (closes #168)